+++ analysis/08_timeToEvent.R
@@ -7,7 +7,7 @@ options(connectionObserver = NULL)
 
 
 ## Set variables ----------------------
-configBlock <- "mrktscan"
+configBlock <- "optum"
 outputFolder <- here::here("output", "08_timeToInitialTreatment", configBlock)
 
 cohortsToCreate <- readr::read_csv(here::here("output", "01_buildCohorts", configBlock, "cohortManifest.csv"),
@@ -22,7 +22,7 @@ startSnowflakeSession(con = con, executionSettings = executionSettings)
 
 ### 1. Time To Event Analysis -------------------------
 
-eraCollapseSize <- c(30,60)
+eraCollapseSize <- c(30, 60, 90, 180)
 
 eventType <- c("ingredient", "class3", "class1")
 
@@ -37,15 +37,14 @@ for (i in 1:length(eraCollapseSize)) {
     for(k in 1:length(eventType)) {
 
   ttiDat <- timeToTreatmentDiscontinuationData(executionSettings = executionSettings,
-                                       eraCollapseSize = eraCollapseSize[i],
-                                       eventType = eventType[k],
-                                       targetCohorts = targetCohorts[j,],
-                                       outputFolder = outputFolder)
+                                                eraCollapseSize = eraCollapseSize[i],
+                                                eventType = eventType[k],
+                                                targetCohorts = targetCohorts[j,],
+                                                outputFolder = outputFolder)
 
     }
   }
 }
 
 
-View(ttiDat$total)